[ALPS05769886] cmdq: fix coverity issue with invalid type in argument
fix coverity issue with invalid type in argument

MTK-Commit-Id: 4d935fbc3b1d68a04502de05241f7016b0e57ea0

CR-Id: ALPS05769886
Feature: [Module]CMDQ Driver

+++ drivers/misc/mediatek/cmdq/v3/cmdq_sec.c
@@ -1992,7 +1992,7 @@ static void cmdq_sec_exec_task_async_impl(struct work_struct *work_item)
 
 			CMDQ_ERR(
 				"gce: err:%d task:%p pkt:%p thread:%u task_cnt:%u wait_cookie:%u next_cookie:%u",
-				(unsigned long) err, task, task->handle->pkt,
+				err, task, task->handle->pkt,
 				task->thread->idx, task->thread->task_cnt,
 				task->thread->wait_cookie, task->thread->next_cookie);
 			spin_unlock_irqrestore(&task->thread->chan->lock, flags);
